Adjust changes requested
